+++ release_notes.md
@@ -1,5 +1,6 @@
 ###In Development
   - [#143](https://github.com/MehdiK/Humanizer/pull/143): Added Russian translation for future DateTime, TimeSpan and Now
+  - [#144](https://github.com/MehdiK/Humanizer/pull/144): Danish localization (strings, tests)
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.18.1...master)