+++ release_notes.md
@@ -1,5 +1,7 @@
 ###In Development
 
+- [#144](https://github.com/MehdiK/Humanizer/pull/144): Danish localization (strings, tests)
+
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.18.1...master)
 
 ###v1.18.1 - 2014-04-09