+++ release_notes.md
@@ -1,5 +1,5 @@
 ###In Development
-
+- [#146](https://github.com/MehdiK/Humanizer/pull/146): Added Spanish translation for future DateTime, TimeSpan and Now
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.18.1...master)
 
 ###v1.18.1 - 2014-04-09