+++ release_notes.md
@@ -1,10 +1,14 @@
 ###In Development
+
+[Commits](https://github.com/MehdiK/Humanizer/compare/v1.19.1...master)
+
+###v1.19.1 - 2014-04-10
+  - [#149](https://github.com/MehdiK/Humanizer/pull/149): Improved & refactored number to words localisation 
   - [#143](https://github.com/MehdiK/Humanizer/pull/143): Added Russian translation for future DateTime, TimeSpan and Now
-  - [#144](https://github.com/MehdiK/Humanizer/pull/144): Danish localization (strings, tests)
-  - [#149](https://github.com/MehdiK/Humanizer/pull/149): Number to words localisation improvements
+  - [#144](https://github.com/MehdiK/Humanizer/pull/144): Added Danish localization (strings, tests)
   - [#146](https://github.com/MehdiK/Humanizer/pull/146): Added Spanish translation for future DateTime, TimeSpan and Now
 
-[Commits](https://github.com/MehdiK/Humanizer/compare/v1.18.1...master)
+[Commits](https://github.com/MehdiK/Humanizer/compare/v1.18.1...v1.19.1)
 
 ###v1.18.1 - 2014-04-09
   - [#137](https://github.com/MehdiK/Humanizer/pull/137): Fixed grammar error in `nb-NO` resource file & added missing Norwegian resource strings (mainly `DateHumanize_*FromNow`)